Imported using git-ubuntu import.

Publish parent: ec494ad6e1aa04042eae15d9e732ac7e699a88b1
Changelog parent: d588dbd1f00363b42302dad9bceb42b50d69d31a
Unapplied parent: 083083a23976beafcf425407d43b7565d282bab1

+++ sysdeps/unix/sysv/linux/aarch64/sys/procfs.h
@@ -1,4 +1,4 @@
-/* Copyright (C) 1996-2016 Free Software Foundation, Inc.
+/* Copyright (C) 1996-2015 Free Software Foundation, Inc.
 
    This file is part of the GNU C Library.
 
@@ -91,8 +91,8 @@ struct elf_prpsinfo
     char pr_zomb;			/* Zombie.  */
     char pr_nice;			/* Nice val.  */
     unsigned long int pr_flag;		/* Flags.  */
-    unsigned int pr_uid;
-    unsigned int pr_gid;
+    unsigned short int pr_uid;
+    unsigned short int pr_gid;
     int pr_pid, pr_ppid, pr_pgrp, pr_sid;
     /* Lots missing */
     char pr_fname[16];			/* Filename of executable.  */
